fix botched docs merge
