--- Sources/HVCaptureSDK/HVCaptureSDK.docc/CaptureSDK.docc/CaptureSDK.md
@@ -1,30 +1,0 @@
-# ``_HVCameraExterior``
-
-The HOVER capture experience
-
-@Metadata {
-    @PageImage(
-        purpose: icon, 
-        source: "hvcamera-icon", 
-        alt: "An icon representing the HVCamera framework.")
-    @PageColor(blue)
-}
-
-## Overview
-
-The HOVER Capture SDK provides on-device functionality to facilitate HOVER 3D model creation. Specifically, it provides:
-
-- Sensor interaction for data acquisition 
-- On-device data processing and transformation
-- Visual elements and user guidance to guide data acquisition and improve captured data
-- On-device local persistence
-- Robust data uploads to the HOVER backend
-
-Once the data has been captured and uploaded from the SDK, the HOVER backend will begin the 3D model processing.   
-
-## Essentials
-
-@Links(visualStyle: compactGrid) {
-    - <doc:GettingStarted>
-    - <doc:BasicSDK>
-}
