--- Sources/HVCaptureSDK/HVCaptureSDK.docc/CaptureSDK.docc/Resources/code-files/tutorial-sdk-code-01.swift
@@ -1,14 +1,0 @@
-import SwiftUI
-import HVCaptureSDK
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
