--- Sources/HVCaptureSDK/HVCaptureSDK.docc/CaptureSDK.docc/Resources/code-files/tutorial-sdk-code-02.swift
@@ -1,17 +1,0 @@
-import SwiftUI
-import HVCaptureSDK
-
-@main
-struct BinarySDKExampleApp: App {
-
-    init() {
-        let params = HVCameraExteriorParameters()
-        HVCameraExterior.sharedInstance.initialize(parameters: params)
-    }
-
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
