--- Sources/HVCaptureSDK/HVCaptureSDK.docc/CaptureSDK.docc/Resources/code-files/tutorial-sdk-code-03.swift
@@ -1,30 +1,0 @@
-import SwiftUI
-import HVCaptureSDK
-
-struct ContentView: View {
-
-    var jobInfo: CaptureJobInformation
-    var settings: HVCameraSettings
-
-    init() {
-        let jobIdentifier = JobIdentifier(jobID: 42)
-        jobInfo = CaptureJobInformation(
-            firstTimeUser: true,
-            identifier: jobIdentifier,
-            uploadSecret: "?")
-        settings = HVCameraSettings()
-    }
-
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        Button("Start Capture") {
-
-        }
-        .padding()
-    }
-}
