+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/code-files/tutorial-hvtheme-code-06.swift
