+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/code-files/tutorial-hvtheme-code-07.swift
