+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/code-files/tutorial-sdk-code-05.swift
@@ -9,7 +9,11 @@ struct ContentView: View {
     @State var captureTask: Task<Void, Never>?
 
     init() {
-        jobInfo = CaptureJobInformation(firstTimeUser: true, jobID: 42, clientIdentifier: "hello", userEmail: "world", uploadSecret: "?", isTestJob: true)
+        let jobIdentifier = JobIdentifier(jobID: 42)
+        jobInfo = CaptureJobInformation(
+            firstTimeUser: true,
+            identifier: jobIdentifier,
+            uploadSecret: "?")
         settings = HVCameraSettings()
     }
 
@@ -21,9 +25,10 @@ struct ContentView: View {
             Text("Hello, world!")
         }
         Button("Start Capture") {
-            captureTask = Task {
+            Task {
                 do {
                     try await HVCameraExterior.sharedInstance.startCaptureSession(settings: settings, info: jobInfo)
+                    try await HVCameraExterior.sharedInstance.startCaptureFlow()
                 } catch let error as HVSessionError {
                     // maybe handle our known errors here
                     switch error.kind {
