+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/hover-graph-logo.png
