+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/tutorial-s01-images/tutorial-s01-01.png
