+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/tutorial-s02-images/t2-s4-gridscreen.png
