+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/tutorial-s02-images/t2-s5-text.gif
