+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/tutorial-s02-images/t2p1-button-types.png
