+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Resources/tutorial-s02-images/tutorial-s02-default.png
