+++ Sources/HVCaptureSDK/HVCaptureSDK.docc/Tutorials/Tutorials.tutorial
@@ -1,28 +1,25 @@
 @Tutorials(name: "HOVER Capture SDK") {
     
-    @Metadata {
-        @PageImage(
-            purpose: card, 
-            source: "hvcamera-icon", 
-            alt: "A stylized house.")
-        @PageImage(
-            purpose: icon, 
-            source: "hover-logo", 
-            alt: "An icon representing the HOVER logo framework.")
-        @PageColor(blue)
-    }
-    
     @Intro(title: "Meet the HOVER Capture SDK") {
-        Embed the HOVER Capture experience directly into your application. Get started with the HOVER Capture SDK by building a demo application.  
+        Embed the HOVER Capture experience directly into your application. 
+        Get started with the HOVER Capture SDK by building a demo application.  
         
         @Image(source: hvcamera-icon, alt: "HOVER Capture SDK logo")
     }
     
     @Chapter(name: "Capture SDK Essentials") {
-        Embed the vanilla capture experience in an iOS application
+        Integrate the HOVER Capture SDK in an iOS application
         
         @Image(source: hvcamera-icon, alt: "HOVER Capture SDK logo")
         
         @TutorialReference(tutorial: "doc:BasicSDK")
+    }
+    
+    @Chapter(name: "Customizing Capture SDK Appearance") {
+        Customize aspects of the Capture SDK UI        
+        
+        @Image(source: hvcamera-icon, alt: "HOVER Capture SDK logo")
+        
+        @TutorialReference(tutorial: "doc:HVThemeCustomization")
     }
 }
